@GetMapping("/series/{seriesId}/seasons") @GetMapping("/series/{seriesId}/seasons/{seasonNumber}")

+++ src/main/java/com/netflix/repository/ChaptersRepository.java
@@ -1,14 +1,19 @@
 package com.netflix.repository;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.netflix.model.Categories;
 import com.netflix.model.Chapters;
+import com.netflix.model.Seasons;
+import com.netflix.model.TvShows;
 
 @Repository
 @Qualifier("ChaptersRepository")
-public interface ChaptersRepository  extends JpaRepository<Chapters, Integer>{
-
+public interface ChaptersRepository  extends JpaRepository<Chapters, Long>{
+	
+	
+	
 }
